Move usefull types to views
